fill database in viewmodel
