+++ app/src/main/java/com/lightricks/feedexercise/data/Models.kt
@@ -9,29 +9,9 @@ import com.squareup.moshi.*
  * The rest of the properties are left out for brevity.
  */
 
-@JsonClass(generateAdapter = true)
 data class FeedItem(
     val id: String, // Kotlin note: "val" means read-only value.
-    @Json(name = "templateThumbnailURI") @ThumbnailUrl val thumbnailUrl: String,
+    val thumbnailUrl: String,
     val isPremium: Boolean
 )
 
-
-@Retention(AnnotationRetention.RUNTIME)
-@JsonQualifier
-annotation class ThumbnailUrl
-internal class ThumbnailUrlAdapter {
-
-    private val BASE_URL = "https://assets.swishvideoapp.com/Android/demo/catalog/thumbnails/"
-
-    @ToJson
-    fun toJson(@ThumbnailUrl thumbnailUrl: String): String {
-        return thumbnailUrl
-    }
-
-    @FromJson
-    @ThumbnailUrl
-    fun fromJson(thumbnailUrl: String): String {
-        return "$BASE_URL$thumbnailUrl"
-    }
-}
