+++ app/src/main/java/com/lightricks/feedexercise/database/Entities.kt
@@ -3,7 +3,6 @@ package com.lightricks.feedexercise.database
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.lightricks.feedexercise.data.ThumbnailUrl
 import com.squareup.moshi.Json
 
 /**
