+++ app/src/main/java/com/lightricks/feedexercise/network/DataTransferObjects.kt
@@ -1,7 +1,7 @@
 package com.lightricks.feedexercise.network
 
 import com.lightricks.feedexercise.data.FeedItem
-import com.squareup.moshi.JsonClass
+import com.squareup.moshi.*
 
 /**
  * todo: add Data Transfer Object data class(es) here
@@ -10,5 +10,33 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class GetFeedResponse(
-    val templatesMetadata : List<FeedItem>
+    val templatesMetadata : List<FeedItemDTO>
 )
+
+
+@JsonClass(generateAdapter = true)
+data class FeedItemDTO(
+    val id: String, // Kotlin note: "val" means read-only value.
+    @Json(name = "templateThumbnailURI") @ThumbnailUrl val thumbnailUrl: String,
+    val isPremium: Boolean
+)
+
+
+@Retention(AnnotationRetention.RUNTIME)
+@JsonQualifier
+annotation class ThumbnailUrl
+internal class ThumbnailUrlAdapter {
+
+    private val BASE_URL = "https://assets.swishvideoapp.com/Android/demo/catalog/thumbnails/"
+
+    @ToJson
+    fun toJson(@ThumbnailUrl thumbnailUrl: String): String {
+        return thumbnailUrl
+    }
+
+    @FromJson
+    @ThumbnailUrl
+    fun fromJson(thumbnailUrl: String): String {
+        return "$BASE_URL$thumbnailUrl"
+    }
+}
